+++ protected/humhub/modules/admin/components/Controller.php
@@ -52,15 +52,21 @@ class Controller extends \humhub\components\Controller
             'acl' => [
                 'class' => AccessControl::className(),
                 'adminOnly' => $this->adminOnly,
-                'rules' => static::getAccessRules()
+                'rules' => $this->getAccessRules()
             ]
         ];
     }
 
-    public static function getAccessRules()
+    /**
+     * Returns access rules for the standard access control behavior
+     * 
+     * @see AccessControl
+     * @return array the access permissions
+     */
+    public function getAccessRules()
     {
-        // Workaround for module configuration actions
-        if (Yii::$app->controller->module->id != 'admin') {
+        // Use by default ManageModule permission, if method is not overwritten by custom module
+        if ($this->module->id != 'admin') {
             return [
                 ['permissions' => \humhub\modules\admin\permissions\ManageModules::className()]
             ];